+++ src/main/java/org/m2ci/msp/jtgt/Annotation.java
@@ -1,5 +1,8 @@
 package org.m2ci.msp.jtgt;
 
+import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
+import org.apache.commons.lang3.builder.ToStringStyle;
+import org.json.JSONObject;
 
 /**
  * TextGrid annotation wrapper class
@@ -103,6 +106,11 @@ public abstract class Annotation {
         return ((getStart() == ((Annotation) o).getStart()) &&
                 (getEnd() == ((Annotation) o).getEnd())  &&
                 getText().equals(((Annotation) o).getText()));
+    }
+
+    public String toString() {
+        String jsonStr = new ReflectionToStringBuilder(this, ToStringStyle.JSON_STYLE).toString();
+        return new JSONObject(jsonStr).toString(4);
     }
 }